+++ Source/VK/QueueSemaphoreVK.cpp
@@ -50,7 +50,7 @@ Result QueueSemaphoreVK::Create(void* vkSemaphore)
 
 inline void QueueSemaphoreVK::SetDebugName(const char* name)
 {
-    m_Device.SetDebugNameToTrivialObject(VK_OBJECT_TYPE_SEMAPHORE, m_Handle, name);
+    m_Device.SetDebugNameToTrivialObject(VK_OBJECT_TYPE_SEMAPHORE, (uint64_t)m_Handle, name);
 }
 
 #include "QueueSemaphoreVK.hpp"
